requirements.txt
asn1crypto==1.5.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
filelock==3.13.4
idna==3.7
packaging==24.0
platformdirs==4.2.1
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
snowflake-connector-python==3.9.1
sortedcontainers==2.4.0
tomlkit==0.12.4
typing_extensions==4.11.0
urllib3==2.2.1
